+++ R/module-pwd.R
@@ -127,12 +127,21 @@ pwd_server <- function(input, output, session, user, update_pwd, validate_pwd = 
   observeEvent(input$update_pwd, {
     password$relog <- NULL
     removeUI(selector = jns("msg_pwd"))
+  
     if (!identical(input$pwd_one, input$pwd_two)) {
       insertUI(
         selector = jns("result_pwd"),
         ui = tags$div(
           id = ns("msg_pwd"), class = "alert alert-danger",
           icon("exclamation-triangle"), lan()$get("The two passwords are different")
+        )
+      )
+    } else if (!check_new_pwd(user$user, input$pwd_one)) {
+      insertUI(
+        selector = jns("result_pwd"),
+        ui = tags$div(
+          id = ns("msg_pwd"), class = "alert alert-danger",
+          icon("exclamation-triangle"), lan()$get("New password cannot be the same as old")
         )
       )
     } else {